+++ HISTORY.md
@@ -53,6 +53,7 @@
 * ported multiline library to new (custom input slot) format
 * new "text costumes" library for generating costumes from letters or words of text
 * took out "b block" costume from catalog
+* added microphone "resolution" concept governing "bins" (buffer / bin sizes)
 
 ### 2019-03-10
 * Objects, Blocks, Threads: added microphone note and pitch detection